Phase 6: handler template methods, NPC partial→standalone, IRecordDisplayRenderer, warning threshold tuning
- Add ParseRecordList<T> and ParseAccessorOnly<T> to RecordHandlerBase, migrate 25+ handler methods
- Convert NpcExportSceneBuilder and NpcRenderHelpers partial classes to standalone classes
- Extract IRecordDisplayRenderer interface + 15 adapters for ShowCommand coupling reduction
- Tune .editorconfig: move S3776/CA1502/S1200/IDE0060 to suggestion (domain-inherent complexity)
- Net result: 1,305 → 20 build warnings (-98.5%), -760 lines

+++ src/FalloutXbox360Utils/Core/RuntimeBuffer/RuntimeBufferPointerAnalyzer.cs
@@ -205,7 +205,7 @@ internal sealed class RuntimeBufferPointerAnalyzer
                 continue;
             }
 
-            var regionOffset = (long)alignDelta;
+            var regionOffset = alignDelta;
             while (regionOffset + 4 <= region.Size)
             {
                 var remaining = region.Size - regionOffset;
